ENH: Addition of options to faciliate development with docker
run.sh does not remove the container by default. If one wants to remove the container,
one has to specify the new option '-d'. The container matching the container name that
is hard-coded in the script (i.e. dream3d) will be removed before docker is started,
and upon exit.
If one interrupts docker, and starts it again, the default behavior is to keep using
the container that was used previously. One can give a container ID with the option
'-s' to start a specific container.
Additionally, one can start 'bash' in the running container with the new option '-e'.
This is convenient if one wants to perform actions within the container.