ENH: Clean start or restart of containers, depending if they exist or not
When starting the 'dream3d' container using run.sh, the script ensures
that the container does not already exist. If it exists, and is already
running, the script will cleanly exit. If it exists and the container
is not running: 1) if '-d' is specified, the container is removed,
and a clean new container is starter; 2) if '-d' is not specifier,
the existing container is restarted.
The option '-s' to select which container to start has been removed. It
is always the container named 'dream3d' (hard-coded) that is started.
option is not specified, it will start